Refactor scrapers into portal packages
