+++ app/infrastructure/bumeran/scraper.py
@@ -274,21 +274,4 @@ class BumeranScraper:
             return False
 
 
-def scrap_jobs_bumeran(
-    *,
-    query: str = "",
-    location: str = "",
-    max_pages: int | None = None,
-    job_id: str | None = None,
-    headless: bool = True
-) -> List[Dict[str, Any]]:
-    with sync_playwright() as pw:
-        browser = pw.chromium.launch(headless=False, args=["--start-maximized"])
-        try:
-            scraper = BumeranScraper(
-                browser, query=query, location=location, max_pages=max_pages, job_id=job_id
-            )
-            return scraper.run()
-        finally:
-            browser.close()
