+++ app/infrastructure/computrabajo/cli.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+import argparse
+import json
+import pandas as pd
+
+from . import scrape_computrabajo
+
+
+def main() -> None:
+    parser = argparse.ArgumentParser(description="Scrapea Computrabajo")
+    parser.add_argument("--categoria", required=True)
+    parser.add_argument("--lugar", required=True)
+    parser.add_argument("--pages", type=int, default=None)
+    args = parser.parse_args()
+
+    resultados = scrape_computrabajo(
+        categoria=args.categoria,
+        lugar=args.lugar,
+        max_pages=args.pages,
+    )
+
+    if resultados:
+        df = pd.DataFrame(resultados)
+        df.to_excel("computrabajo.xlsx", index=False)
+        with open("computrabajo.json", "w", encoding="utf-8") as f:
+            json.dump(resultados, f, ensure_ascii=False, indent=2)
+        print(f"[SUCCESS] {len(resultados)} registros exportados.")
+    else:
+        print("[INFO] No se extrajeron datos.")
+
+
+if __name__ == "__main__":
+    main()
